app-python/requirements.txt
prometheus_client==0.20.0
pyaml==24.7.0
requests==2.32.3
